CHANGE: move tracing module under base module
